# docs/api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AI 文章生成后端：用户/会话、文章三阶段流水线（SSE 进度）、配图、Agent 日志；开发态 VIP Mock 支付与管理端升降 VIP。",
        "title": "Wood Passage Creator API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/admin/payment/list": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "全站支付记录列表，仅管理员；支持 pageNum/pageSize/status/userId/productType 查询参数。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-payment"
                ],
                "summary": "分页查询支付记录（管理员）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码，默认 1",
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数，默认 10，最大 100",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "PENDING",
                            "SUCCEEDED",
                            "FAILED",
                            "REFUNDED"
                        ],
                        "type": "string",
                        "description": "状态筛选",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "用户 ID",
                        "name": "userId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "产品类型，如 VIP_PERMANENT",
                        "name": "productType",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/wood-passage-creator_internal_app_payment.RecordListData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
        "/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "description": "登录信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_app_auth.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/wood-passage-creator_internal_module_user.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误/账号或密码错误",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "登出",
                "responses": {
                    "200": {
                        "description": "成功（data 一般为 null）",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "根据 session 从数据库加载最新用户信息（角色/配额等），用于前端刷新登录态。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "获取当前登录用户",
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/wood-passage-creator_internal_module_user.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "注册",
                "parameters": [
                    {
                        "description": "注册信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_app_auth.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/wood-passage-creator_internal_module_user.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误/账号冲突",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/payment/list": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment"
                ],
                "summary": "我的支付记录",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码，默认 1",
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数，默认 10，最大 100",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "PENDING",
                            "SUCCEEDED",
                            "FAILED",
                            "REFUNDED"
                        ],
                        "type": "string",
                        "description": "状态筛选",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "产品类型",
                        "name": "productType",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/wood-passage-creator_internal_app_payment.RecordListData"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/payment/vip/mock-complete": {
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "将 PENDING 会话标为 SUCCEEDED 并为对应用户开通 VIP（幂等）。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment"
                ],
                "summary": "[Dev] Mock 支付成功回调",
                "parameters": [
                    {
                        "description": "sessionId",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_app_payment.MockCompleteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/wood-passage-creator_internal_module_payment.MockCompleteResult"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/payment/vip/mock-session": {
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "不连接 Stripe；写入 PENDING 支付记录并返回假 checkoutUrl。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment"
                ],
                "summary": "[Dev] 创建 Mock VIP 支付会话",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/wood-passage-creator_internal_module_payment.MockSessionResult"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "wood-passage-creator_internal_app_auth.LoginRequest": {
            "type": "object",
            "required": [
                "userAccount",
                "userPassword"
            ],
            "properties": {
                "userAccount": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 3
                },
                "userPassword": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 6
                }
            }
        },
        "wood-passage-creator_internal_app_auth.RegisterRequest": {
            "type": "object",
            "required": [
                "userAccount",
                "userPassword"
            ],
            "properties": {
                "userAccount": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 3
                },
                "userAvatar": {
                    "type": "string",
                    "maxLength": 1024
                },
                "userName": {
                    "type": "string",
                    "maxLength": 256,
                    "minLength": 1
                },
                "userPassword": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 6
                },
                "userProfile": {
                    "type": "string",
                    "maxLength": 512
                }
            }
        },
        "wood-passage-creator_internal_app_payment.MockCompleteRequest": {
            "type": "object",
            "required": [
                "sessionId"
            ],
            "properties": {
                "sessionId": {
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 8
                }
            }
        },
        "wood-passage-creator_internal_app_payment.RecordListData": {
            "type": "object",
            "properties": {
                "pageNum": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "records": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/wood-passage-creator_internal_module_payment.Record"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "wood-passage-creator_internal_module_payment.MockCompleteResult": {
            "type": "object",
            "properties": {
                "isVip": {
                    "type": "boolean"
                },
                "record": {
                    "$ref": "#/definitions/wood-passage-creator_internal_module_payment.Record"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "wood-passage-creator_internal_module_payment.MockSessionResult": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "checkoutUrl": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "productType": {
                    "type": "string"
                },
                "sessionId": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/wood-passage-creator_internal_module_payment.RecordStatus"
                }
            }
        },
        "wood-passage-creator_internal_module_payment.Record": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "createTime": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "productType": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/wood-passage-creator_internal_module_payment.RecordStatus"
                },
                "stripePaymentIntentId": {
                    "type": "string"
                },
                "stripeSessionId": {
                    "type": "string"
                },
                "updateTime": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "wood-passage-creator_internal_module_payment.RecordStatus": {
            "type": "string",
            "enum": [
                "PENDING",
                "SUCCEEDED",
                "FAILED",
                "REFUNDED"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusSucceeded",
                "StatusFailed",
                "StatusRefunded"
            ]
        },
...
        "wood-passage-creator_internal_module_user.User": {
            "type": "object",
            "properties": {
                "createTime": {
                    "type": "string"
                },
                "editTime": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "quota": {
                    "type": "integer"
                },
                "updateTime": {
                    "type": "string"
                },
                "userAccount": {
                    "type": "string"
                },
                "userAvatar": {
                    "type": "string"
                },
                "userName": {
                    "type": "string"
                },
                "userProfile": {
                    "type": "string"
                },
                "userRole": {
                    "$ref": "#/definitions/wood-passage-creator_internal_module_user.UserRole"
                },
                "vipTime": {
                    "type": "string"
                }
            }
        },
        "wood-passage-creator_internal_module_user.UserRole": {
            "type": "string",
            "enum": [
                "user",
                "admin",
                "vip"
            ],
            "x-enum-varnames": [
                "RoleUser",
                "RoleAdmin",
                "RoleVIP"
            ]
        },
        "wood-passage-creator_internal_pkg_response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "SessionAuth": {
            "description": "Session cookie authentication. Example: session=your-session-id",
            "type": "apiKey",
            "name": "Cookie",
            "in": "header"
        }
    }
}